refactor: Modify to final class, naming, access control, magic string
- Modify Item -> koreaItems
- Modify storeUIView -> stackViewItemsArray
- Set setMainView method for view setting
